Creación de hook personalizado useUsers, que contiene manipulación de distintos estados que antes estaban en UsersApp
