+++ src/UsersApp.jsx
@@ -1,54 +1,10 @@
-import React, { useEffect, useReducer, useState } from 'react';
 import UsersList from './Components/UsersList';
 import UserForm from './Components/UserForm';
-import { usersReducer } from './reducers/usersReducer';
+import useUsers from './hooks/useUsers';
 
-const useListSession = JSON.parse(sessionStorage.getItem("usersList")) || [];
-const form = {
-  id: 0,
-  username: "",
-  password: "",
-  email: ""
-}
 const UsersApp = () => {
 
-  const [usersList, dispatch] = useReducer(usersReducer, useListSession)
-  const [formUpdate, setFormUpdate] = useState(form)
-
-  const handlerUser = (infoUser) => {
-    // Verifica si el usuario ya existe por ID
-    const exists = usersList.some(user => user.id === infoUser.id);
-    if (exists) {
-      dispatch({
-        type: 'UpdateUser',
-        payload: infoUser
-      });
-    } else {
-      dispatch({
-        type: 'AddUser',
-        payload: infoUser
-      });
-    }
-  }
-
-  useEffect(() => {
-    sessionStorage.setItem("usersList", JSON.stringify(usersList))
-  }, [usersList])
-
-  const handlerDeleteUser = (id) => {
-    console.log(id)
-    dispatch(
-      {
-        type: 'RemoveUser',
-        payload: id
-      }
-    )
-  }
-
-  const handlerUserForm = (infoUserUpdate) => {
-    setFormUpdate({ ...infoUserUpdate })
-  }
-
+  const {form, usersList, formUpdate, handlerUser, handlerDeleteUser, handlerUserForm} = useUsers();
 
   return (
     <div className='container my-4'>
